extract controllers for easy repl access
