+++ server/routes/visit.js
@@ -5,50 +5,9 @@
 'use strict';
 
 var Joi = require('joi');
-var Boom = require('boom');
 
 var log = require('../logger')('server.routes.visit');
-var visit = require('../models/visit');
-
-var visitController = {
-  get: function (request, reply) {
-    var fxaId = request.auth.credentials;
-    visit.get(fxaId, request.params.visitId, function(err, result) {
-      if (err) {
-        log.warn(err);
-        return reply(Boom.create(500));
-      }
-      if (!result) {
-        return reply(Boom.create(404, 'Visit not found'));
-      } else {
-        reply(result);
-      }
-    });
-  },
-  put: function(request, reply) {
-    var fxaId = request.auth.credentials;
-    var visitId = request.params.visitId;
-    var p = request.payload;
-    visit.update(fxaId, visitId, p.visitedAt, p.url, p.title, function(err, result) {
-      if (err) {
-        log.warn(err);
-        return reply(Boom.create(500));
-      }
-      // return the visit so backbone can update the model
-      reply(result);
-    });
-  },
-  delete: function (request, reply) {
-    var fxaId = request.auth.credentials;
-    visit.delete(fxaId, request.params.visitId, function(err) {
-      if (err) {
-        log.warn(err);
-        return reply(Boom.create(500));
-      }
-      reply();
-    });
-  }
-};
+var visitController = require('../controllers/visit');
 
 var visitRoutes = [{
   method: 'GET',
